+++ inst/PEMcollectr/server.R
@@ -1,10 +1,14 @@
 function(input, output, session) {
   sfObject <- geomUploadServer('geomUpload')
-  submitIds <- validateTableServer('validateTable', sfObject = sfObject)
+  submitIds <- validateTableServer('validateTable', sfObject = sfObject,
+    success = dbWrite, con = con)
   dbWrite <- dbWriteServer('dbWrite',
     sfObject = shiny::reactive({
       sfObject()[sfObject()[['transect_id']] %in% submitIds(), ]}),
-    tableName = DBI::SQL('staging.field_data_points'),
+    tableName = shiny::reactive({
+      staging_tables()[[guess_geometry_type(sfObject())]]
+      }),
     con = con)
-  validatePairsServer('validatePairs', sfObject = sfObject)
+  validatePairsServer('validatePairs', sfObject = sfObject, success = dbWrite,
+    con = con)
 }